+++ backend/back/views_user.py
@@ -31,21 +31,26 @@ views_user = Blueprint('views_user', __name__)
 
 #     return user_schema.jsonify(new_user)
 
-@views_user.route('/all_users', methods=['GET'])
+@views_user.route('/', methods=['GET'])
+@jwt_required()
 def get_users():
     all_users = User.query.all()
+    if len(all_users) == 0:
+        return jsonify("Users not found!"), 404
     result = users_schema.dump(all_users)
     return jsonify(result)
 
-@views_user.route('/user/<id>', methods=['GET'])
-def get_user(id):
-    user = User.query.get(id)
+@views_user.route('/<uuid:user_id>', methods=['GET'])
+@jwt_required()
+def get_user(user_id):
+    user = User.query.get(user_id)
+    if user is None:
+        return jsonify("User not found!"), 404
     return user_schema.jsonify(user)
 
-@views_user.route('/user/<id>', methods=['PUT'])
-def update_user(id):
-
-    user = User.query.get(id)
+@views_user.route('/<uuid:user_id>', methods=['PUT'])
+@jwt_required()
+def update_user(user_id):
 
     active = request.json['active']
     address = request.json['address']
@@ -58,27 +63,41 @@ def update_user(id):
     role = request.json['role']
     surname = request.json['surname']
 
-    user.active = active
-    user.address = address
-    user.authorize_date = authorize_date
-    user.birth_date = birth_date
-    user.date_created = user.date_created
-    user.date_of_last_login = user.date_of_last_login
-    user.email = email
-    user.end_authorize_date = end_authorize_date
-    user.name = name
-    user.password = user.password
-    user.phone = phone
-    user.role = role
-    user.surname = surname
+    try:
+        user = User.query.get(user_id)
+        if user is None:
+            return jsonify("User not found"), 404
 
-    db.session.commit()
+        user.active = active
+        user.address = address
+        user.authorize_date = authorize_date
+        user.birth_date = birth_date
+        user.date_created = user.date_created
+        user.date_of_last_login = user.date_of_last_login
+        user.email = email
+        user.end_authorize_date = end_authorize_date
+        user.name = name
+        user.password = user.password
+        user.phone = phone
+        user.role = role
+        user.surname = surname
 
-    return user_schema.jsonify(user)
+        db.session.commit()
 
-@views_user.route('/user/<id>', methods=["DELETE"])
-def delete_user(id):
-    user = User.query.get(id)
-    db.session.delete(user)
-    db.session.commit()
-    return jsonify("User " + id + " was succesfully deleted!")
+        return user_schema.jsonify(user)
+    except:
+        return jsonify("Failure in modifying an user"), 422
+
+@views_user.route('/<uuid:user_id>', methods=["DELETE"])
+@jwt_required()
+def delete_user(user_id):
+    try:
+        user = User.query.get(user_id)
+        if user is None:
+            return jsonify("User not found"), 404
+        db.session.delete(user)
+        db.session.commit()
+        return jsonify("User " + str(user_id) + " was succesfully deleted!")
+    except:
+        return jsonify("Faliure in deleting an user"), 500
+
